chemDB/requirements.txt
pymysql
pandas
sqlalchemy
selenium
requests
beautifulsoup4
lxml
html5lib
zeep
python-dotenv
tabula
odfpy